third_party/blink/web_tests/platform/generic/fast/dom/implementation-createHTMLDocument-expected.txt

This page tests the DOM createHTMLDocument method.

If the test passes, you'll see a series of 'PASS' messages below.

[document with title 'title']
PASS: doc.title should be title and is.
PASS: doc.getElementsByTagName('html').length should be 1 and is.
PASS: doc.getElementsByTagName('head').length should be 1 and is.
PASS: doc.getElementsByTagName('title').length should be 1 and is.
PASS: doc.getElementsByTagName('body').length should be 1 and is.
[document with title '']
PASS: doc.title should be and is.
PASS: doc.getElementsByTagName('title').item(0).firstChild.data should be and is.
[document with null title]
PASS: doc.title should be null and is.
